@@ -1,6 +1,15 @@
+# frozen_string_literal: true
+
 source "https://rubygems.org"
 
-ruby "2.5.1"
+ruby "3.2.2"
 
+gem "mechanize", "~> 2.8.5"
+gem "nokogiri", "~> 1.16.5" # Latest version the platform supports
+gem "rake", "~> 12.3"
+gem "rspec", "~> 3.0"
+gem "rubocop"
 gem "scraperwiki", git: "https://github.com/openaustralia/scraperwiki-ruby.git", branch: "morph_defaults"
-gem "mechanize"
+gem "simplecov", "~> 0.18.0"
+gem "simplecov-console"
+gem "sqlite3", "~> 1.6.3"
